Sistema de Banco de Dados Implementado , Tela de Confirmação e Tela de Historico

+++ lib/model/DadosSalvos.dart
@@ -2,7 +2,7 @@
  
  class DadosSalvos{
   String modelo;
-  double valor;
+  int valor;
   String data;
   String hora;
